move shape functions to derive_globals, add surface area to globals
